+++ src/docx.h
@@ -8,7 +8,7 @@
 int extract_docx_run_start(
         extract_astring_t*  content,
         const char*         font_name,
-        float               font_size,
+        double              font_size,
         int                 bold,
         int                 italic
         );
